refactor: split podman_info into info/hooks_dir/network submodules
The 284-line podman_info.py mixed three independent concerns: parsing
podman info output, discovering OCI hook directories, and computing
slirp4netns/resolv.conf network values.  Promote the module to a
package whose submodules each own one concern:

- ``info.py`` — ``PodmanInfo`` dataclass + ``parse_podman_info`` +
  ``_parse_version`` + ``HOOKS_DIR_PERSIST_VERSION``.
- ``hooks_dir.py`` — global hook directory discovery from
  ``containers.conf`` + ``has_global_hooks`` + ``system_hooks_dir`` +
  ``global_hooks_hint`` + ``USER_HOOKS_DIR``/``HOOK_JSON_FILENAME``.
- ``network.py`` — ``slirp4netns_gateway`` + ``parse_slirp4netns_cidr``
  + ``parse_resolv_conf``.
- ``_conf.py`` (private) — ``_SYSTEM_CONF_PATHS`` +
  ``_user_containers_conf`` shared by ``hooks_dir`` and ``network``.

The package ``__init__.py`` re-exports the same public names so the
``terok_shield`` top-level surface is unchanged.  Internal callers
(hooks/install, hooks/mode, cli/main) and the test suite move to
specific submodule imports for clearer intent.

Tests follow the same split into ``tests/unit/podman_info/``.
