+++ tests/unit/podman_info/test_network.py
@@ -1,0 +1,104 @@
+# SPDX-FileCopyrightText: 2026 Jiri Vyskocil
+# SPDX-License-Identifier: Apache-2.0
+
+"""Tests for ``terok_shield.podman_info.network`` — slirp4netns + resolv.conf."""
+
+from pathlib import Path
+
+import pytest
+
+from terok_shield.podman_info.network import (
+    _parse_network_cmd_options,
+    parse_resolv_conf,
+    parse_slirp4netns_cidr,
+    slirp4netns_gateway,
+)
+
+# ── parse_resolv_conf tests ──────────────────────────────
+
+
+class TestParseResolvConf:
+    """Tests for resolv.conf nameserver parsing."""
+
+    def test_standard_resolv_conf(self) -> None:
+        """Standard resolv.conf with one nameserver."""
+        text = "# Generated by NetworkManager\nnameserver 10.0.2.3\n"
+        assert parse_resolv_conf(text) == "10.0.2.3"
+
+    def test_pasta_dns(self) -> None:
+        """Pasta-generated resolv.conf with link-local DNS."""
+        text = "nameserver 169.254.1.1\n"
+        assert parse_resolv_conf(text) == "169.254.1.1"
+
+    def test_multiple_nameservers_returns_first(self) -> None:
+        """Multiple nameservers → returns the first one."""
+        text = "nameserver 10.0.2.3\nnameserver 8.8.8.8\n"
+        assert parse_resolv_conf(text) == "10.0.2.3"
+
+    def test_empty_file(self) -> None:
+        """Empty file → empty string."""
+        assert parse_resolv_conf("") == ""
+
+    def test_no_nameserver(self) -> None:
+        """File with only comments → empty string."""
+        text = "# comment\nsearch example.com\n"
+        assert parse_resolv_conf(text) == ""
+
+    def test_comments_and_whitespace(self) -> None:
+        """Nameserver after comments and whitespace."""
+        text = "# DNS\n\n  nameserver 10.0.2.3  \n"
+        assert parse_resolv_conf(text) == "10.0.2.3"
+
+
+# ── slirp4netns gateway / CIDR tests ──────────────────
+
+
+class TestSlirp4netnsGateway:
+    """Tests for slirp4netns gateway and CIDR parsing."""
+
+    def test_default_gateway(self) -> None:
+        """Default CIDR produces 10.0.2.2."""
+        assert slirp4netns_gateway() == "10.0.2.2"
+
+    def test_custom_cidr(self) -> None:
+        """Custom CIDR computes base + 2."""
+        assert slirp4netns_gateway("192.168.0.0/24") == "192.168.0.2"
+
+    def test_parse_cidr_from_user_conf(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
+    ) -> None:
+        """User containers.conf with cidr= is picked up."""
+        conf = tmp_path / "containers" / "containers.conf"
+        conf.parent.mkdir()
+        conf.write_text('[engine]\nnetwork_cmd_options = ["cidr=172.16.0.0/24"]\n')
+        monkeypatch.setenv("XDG_CONFIG_HOME", str(tmp_path))
+        monkeypatch.setattr("terok_shield.podman_info.network._SYSTEM_CONF_PATHS", ())
+        assert parse_slirp4netns_cidr() == "172.16.0.0/24"
+
+    def test_parse_cidr_defaults_when_absent(
+        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
+    ) -> None:
+        """Returns default CIDR when no config sets cidr=."""
+        monkeypatch.setenv("XDG_CONFIG_HOME", str(tmp_path / "empty"))
+        monkeypatch.setattr("terok_shield.podman_info.network._SYSTEM_CONF_PATHS", ())
+        assert parse_slirp4netns_cidr() == "10.0.2.0/24"
+
+    def test_parse_network_cmd_options_missing_file(self, tmp_path: Path) -> None:
+        """Missing file returns empty list."""
+        assert _parse_network_cmd_options(tmp_path / "nope.conf") == []
+
+    def test_parse_network_cmd_options_bad_toml(self, tmp_path: Path) -> None:
+        """Corrupt TOML returns empty list."""
+        bad = tmp_path / "bad.conf"
+        bad.write_text("not valid {{{{ toml")
+        assert _parse_network_cmd_options(bad) == []
+
+    def test_parse_network_cmd_options_extracts_list(self, tmp_path: Path) -> None:
+        """Valid config returns the option strings."""
+        conf = tmp_path / "ok.conf"
+        conf.write_text('[engine]\nnetwork_cmd_options = ["cidr=10.1.0.0/16", "mtu=1500"]\n')
+        assert _parse_network_cmd_options(conf) == ["cidr=10.1.0.0/16", "mtu=1500"]
+
+    def test_malformed_cidr_falls_back_to_default(self) -> None:
+        """Malformed CIDR override falls back to default gateway."""
+        assert slirp4netns_gateway("not-a-cidr") == "10.0.2.2"
